Produces a geojson file for NGCA APREF analysis
This script needs to sit in the NGCA folder where 'adjustments' is a sub folder. It will use the adj files for each cluster to simulate a minimally constrained adjustment. It will then use this to produce coordinate comparisons against APREF. The complete archive will be printed to the on geojson file.